% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R.annot.R
\name{add}
\alias{add}
\title{add}
\usage{
add(a, b)

add(a, b)
}
\arguments{
\item{a}{Number one.}

\item{b}{Number two.}
}
\value{
Sum of two number.
}
\description{

}
\examples{
add(a =  , b =  )
library(rPackageTutorial)
add(1, 2)
}
